+++ src/emulator/initEmulators.ts
@@ -7,7 +7,6 @@ import { detectStartCommand } from "./apphosting/developmentServer";
 import { EmulatorLogger } from "./emulatorLogger";
 import { Emulators } from "./types";
 import { Env, maybeGenerateEmulatorYaml } from "../apphosting/config";
-import { detectProjectRoot } from "../detectProjectRoot";
 import { Config } from "../config";
 import { getProjectId } from "../projectUtils";
 import { grantEmailsSecretAccess } from "../apphosting/secrets";
@@ -42,8 +41,7 @@ export const AdditionalInitFns: AdditionalInitFnsType = {
     const projectId = getProjectId(config.options);
     let env: Env[] | null = [];
     try {
-      const projectRoot = detectProjectRoot({ cwd: config.options.cwd }) ?? backendRoot;
-      env = await maybeGenerateEmulatorYaml(projectId, projectRoot);
+      env = await maybeGenerateEmulatorYaml(projectId, backendRoot);
     } catch (e) {
       logger.log("WARN", "failed to export app hosting configs");
     }